refactor: move Regional Prompt Studio slice
# Conflicts:
#	nodes.py
#	tests/fixtures/python_backend_baseline.json
#	tests/test_nodes_module_analyzer.py
