+++ tests/test_python_backend_analyzer.py
@@ -683,9 +683,9 @@ ignored/
 
         self.assertEqual(analyzer.render_json(report), expected_text)
         self.assertEqual(report["schema_version"], 2)
-        self.assertEqual(report["inventory"]["module_count"], 53)
-        self.assertEqual(len(report["registry"]["shipped_python_modules"]), 53)
-        self.assertEqual(len(report["registry"]["runtime_import_closure"]), 52)
+        self.assertEqual(report["inventory"]["module_count"], 55)
+        self.assertEqual(len(report["registry"]["shipped_python_modules"]), 55)
+        self.assertEqual(len(report["registry"]["runtime_import_closure"]), 54)
         self.assertEqual(report["registry"]["missing_internal_imports"], [])
         self.assertEqual(
             report["registry"]["unreachable_shipped_python_modules"],
@@ -721,6 +721,7 @@ ignored/
                 "easyuse_anima/nodes/naia_nodes.py",
                 "easyuse_anima/nodes/prompt_data_nodes.py",
                 "easyuse_anima/nodes/prompt_nodes.py",
+                "easyuse_anima/nodes/regional_nodes.py",
                 "easyuse_anima/nodes/wildcard_nodes.py",
                 "easyuse_anima/profiles/__init__.py",
                 "easyuse_anima/profiles/contract.py",
@@ -731,6 +732,7 @@ ignored/
                 "easyuse_anima/prompt/conditioning.py",
                 "easyuse_anima/prompt/data.py",
                 "easyuse_anima/prompt/fields.py",
+                "easyuse_anima/prompt/regional.py",
             }.issubset(report["registry"]["shipped_python_modules"])
         )
         self.assertIn("nodes.py", report["registry"]["shipped_python_modules"])
@@ -756,6 +758,7 @@ ignored/
                 "easyuse_anima/nodes/naia_nodes.py",
                 "easyuse_anima/nodes/prompt_data_nodes.py",
                 "easyuse_anima/nodes/prompt_nodes.py",
+                "easyuse_anima/nodes/regional_nodes.py",
                 "easyuse_anima/nodes/wildcard_nodes.py",
                 "easyuse_anima/prompt/__init__.py",
                 "easyuse_anima/prompt/correction.py",
@@ -763,6 +766,7 @@ ignored/
                 "easyuse_anima/prompt/conditioning.py",
                 "easyuse_anima/prompt/data.py",
                 "easyuse_anima/prompt/fields.py",
+                "easyuse_anima/prompt/regional.py",
                 "easyuse_anima/profiles/__init__.py",
                 "easyuse_anima/profiles/contract.py",
                 "easyuse_anima/profiles/mutation.py",
